Moved lib/ back into src/; Increased eslint scope and ran the linter.
