+++ src/commands/projects/close.js
@@ -22,9 +22,9 @@
  */
 const flow = require('lodash.flow')
 
-const commonYargs = require('../../../lib/common-yargs')
-const printOutput = require('../../../lib/print-output')
-const authenticatedOctokit = require('../../../lib/octokit')
+const commonYargs = require('../../lib/common-yargs')
+const printOutput = require('../../lib/print-output')
+const authenticatedOctokit = require('../../lib/octokit')
 
 /**
  * yargs builder function.
