+++ src/commands/pulls/delete-comment.js
@@ -4,9 +4,9 @@
  * /repos/:owner/:repo/pulls/comments/:comment_id
  */
 const flow = require('lodash.flow')
-const commonYargs = require('../../../lib/common-yargs')
-const printOutput = require('../../../lib/print-output')
-const authenticatedOctokit = require('../../../lib/octokit')
+const commonYargs = require('../../lib/common-yargs')
+const printOutput = require('../../lib/print-output')
+const authenticatedOctokit = require('../../lib/octokit')
 
 /**
  * yargs builder function.
