+++ src/commands/pulls/delete-review-request.js
@@ -5,9 +5,9 @@
  */
 const flow = require('lodash.flow')
 
-const commonYargs = require('../../../lib/common-yargs')
-const printOutput = require('../../../lib/print-output')
-const authenticatedOctokit = require('../../../lib/octokit')
+const commonYargs = require('../../lib/common-yargs')
+const printOutput = require('../../lib/print-output')
+const authenticatedOctokit = require('../../lib/octokit')
 
 /**
  * yargs builder function.
