+++ src/commands/pulls/open.js
@@ -5,9 +5,9 @@
  * /repos/:owner/:repo/pulls/:number
  */
 const flow = require('lodash.flow')
-const commonYargs = require('../../../lib/common-yargs')
-const printOutput = require('../../../lib/print-output')
-const authenticatedOctokit = require('../../../lib/octokit')
+const commonYargs = require('../../lib/common-yargs')
+const printOutput = require('../../lib/print-output')
+const authenticatedOctokit = require('../../lib/octokit')
 
 /**
  * yargs builder function.
