+++ src/lib/print-output.js
